Case Study 1
Using SQL window functions to determine MVM Status for each customer, get the current daily revenue* (the money that the customer is
ingesting into our company) in the monitoring data, and compare it with the previous daily
revenue that will indicate the customer purchasing behavior to get the segment of the customer
whether they are:
a. High Up-normal
b. Up-normal
c. Normal
d. Risk
e. High Risk